refactor(perl-lsp-rs): split hover extraction and tests into submodules (#0000)
